Debug compilation, and syntax errors
