+++ FatFs_SPI/sd_driver/sd_spi.c
@@ -90,7 +90,7 @@ bool sd_spi_transfer(sd_card_t *pSD, const uint8_t *tx, uint8_t *rx,
 
 uint8_t sd_spi_write(sd_card_t *pSD, const uint8_t value) {
     // TRACE_PRINTF("%s\n", __FUNCTION__);
-    u_int8_t received = SPI_FILL_CHAR;
+    uint8_t received = SPI_FILL_CHAR;
 #if 0
     int num = spi_write_read_blocking(pSD->spi->hw_inst, &value, &received, 1);    
     myASSERT(1 == num);
