+++ sensors/tmp117.cpp
@@ -1,12 +1,13 @@
 #include <stdio.h>
 #include <string.h>
-#include "tmp117.h"
 #include "pico/stdlib.h"
 #include "hardware/i2c.h"
 #include "../main.h"
-#include "..misc.h"
+#include "../misc.h"
+#include "tmp117.h"
 
 static int tmp117_addr = 0x48;
+static int DeviceAddress = 0;
 
 //#define I2C_PORT_0  i2c0                Defined in "../main.h"
 
@@ -22,19 +23,22 @@ static int tmp117_addr = 0x48;
 static void write_register16(uint8_t reg, uint8_t data0, uint8_t data1) {
     uint8_t buf[3];
     buf[0] = reg; buf[1] = data0; buf[2] = data1;
-    i2c_write_blocking(I2C_PORT_0, tmp117_addr, buf, 3, true);
+    i2c_write_blocking(I2C_PORT_0, DeviceAddress, buf, 3, true);
 }
 
 static int32_t tmp117_read_raw() {
-    uint8_t buffer[2];
-    uint8_t val = TemperatureRegister | READ_BIT;
-    i2c_write_blocking(I2C_PORT_0, tmp117_addr, &val, 1, true);
-    i2c_read_blocking(I2C_PORT_0, tmp117_addr, buffer, 2, false);
-    int16_t temp = buffer[0] << 8 | buffer[1];
-    return ( -(temp & 0b1000000000000000) | (temp & 0b0111111111111111) );
+    if (DeviceAddress > 0) {
+        uint8_t buffer[2];
+        uint8_t val = TemperatureRegister | READ_BIT;
+        i2c_write_blocking(I2C_PORT_0, DeviceAddress, &val, 1, true);
+        i2c_read_blocking(I2C_PORT_0, DeviceAddress, buffer, 2, false);
+        int16_t temp = buffer[0] << 8 | buffer[1];
+        return ( -(temp & 0b1000000000000000) | (temp & 0b0111111111111111) );
+    }
 }
 
-int initTMP117() {
+void initTMP117()
+{
 
     // setup i2c
     // Set up in "../main.cpp"
@@ -45,12 +49,22 @@ int initTMP117() {
     //gpio_pull_up(SCL_PIN);
 
     // Configuration
-    write_register16(ConfigurationRegister, 0x00, 0x60);
+    uint8_t rxdata;
+	
+    if (i2c_read_blocking(I2C_PORT_0, DeviceAddress = tmp117_addr, &rxdata, 1, false) <= 0)
+	{
+        printf("FAIL (No Device on I2C 0)");
+		DeviceAddress = 0;
+	}
+    if (DeviceAddress > 0){
+        write_register16(ConfigurationRegister, 0x00, 0x60);
+    }
+    return;
 }
 
 void readTMP117(struct STATE *state ) {
-
-    int32_t temperature_raw = tmp117_read_raw();
-    state -> TMP117Temperature = temperature_raw * 7.8125 /1000;
-
+    if (DeviceAddress > 0){
+        int32_t temperature_raw = tmp117_read_raw();
+        state -> TMP117Temperature = temperature_raw * 7.8125 /1000;
+    }
 }
