+++ sensors/tmp117.h
@@ -1,6 +1,6 @@
 #ifndef TMP117_INCLUDED
 #define TMP117_INCLUDED
 
-int initTMP117();
+void initTMP117();
 void readTMP117(struct STATE *s);
 #endif
